# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transform Church Leadership Training Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  clerkId: string;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface UserInput {
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
}

export type RoleUpdateRole = typeof RoleUpdateRole[keyof typeof RoleUpdateRole];


export const RoleUpdateRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
}

export interface Track {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface TrackInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface TrackUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Module {
  id: number;
  trackId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  order: number;
  createdAt: string;
}

export interface TrackWithModules {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  modules: Module[];
}

export interface ModuleInput {
  trackId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  order: number;
}

export interface ModuleUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  order?: number;
}

export interface VideoWithProgress {
  id: number;
  moduleId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  order: number;
  createdAt: string;
  /** @nullable */
  progressPercent?: number | null;
  completed: boolean;
  inQueue: boolean;
}

export interface QuizResult {
  id: number;
  userId: string;
  moduleId: number;
  score: number;
  totalQuestions: number;
  passed: boolean;
  takenAt: string;
}

export interface ModuleDetail {
  id: number;
  trackId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  order: number;
  createdAt: string;
  videos: VideoWithProgress[];
  quizResult?: QuizResult;
  quizUnlocked: boolean;
}

export interface Video {
  id: number;
  moduleId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  order: number;
  createdAt: string;
}

export interface VideoInput {
  moduleId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  order: number;
}

export interface VideoUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  url?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  order?: number;
}

export interface WatchHistoryEntry {
  id: number;
  userId: string;
  videoId: number;
  video: Video;
  progressPercent: number;
  completed: boolean;
  lastWatchedAt: string;
}

export interface WatchProgressInput {
  /**
     * @minimum 0
     * @maximum 100
     */
  progressPercent: number;
  completed?: boolean;
}

export interface QueueItem {
  id: number;
  userId: string;
  videoId: number;
  video: Video;
  addedAt: string;
}

export interface QuizQuestion {
  id: number;
  moduleId: number;
  questionText: string;
  options: string[];
  correctIndex: number;
  order: number;
}

export interface QuizQuestionInput {
  questionText: string;
  options: string[];
  correctIndex: number;
  order: number;
}

export type QuizSubmissionAnswersItem = {
  questionId: number;
  selectedIndex: number;
};

export interface QuizSubmission {
  answers: QuizSubmissionAnswersItem[];
}

export interface Assignment {
  id: number;
  userId: string;
  moduleId: number;
  module: Module;
  assignedBy: string;
  assignedAt: string;
  /** @nullable */
  dueDate?: string | null;
  quizResult?: QuizResult;
}

export interface AssignmentInput {
  userIds: string[];
  moduleId: number;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardSummary {
  assignedModules: Assignment[];
  totalModulesAvailable: number;
  totalModulesCompleted: number;
  totalVideosWatched: number;
  queueCount: number;
  recentActivity: WatchHistoryEntry[];
}

export type UserWithProgressRole = typeof UserWithProgressRole[keyof typeof UserWithProgressRole];


export const UserWithProgressRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface UserWithProgress {
  id: string;
  clerkId: string;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserWithProgressRole;
  createdAt: string;
  quizResults: QuizResult[];
  assignmentCount: number;
}

export type ProgressRowResultsItem = {
  moduleId: number;
  /** @nullable */
  passed?: boolean | null;
  /** @nullable */
  score?: number | null;
};

export interface ProgressRow {
  user: User;
  results: ProgressRowResultsItem[];
}

export interface ProgressMatrix {
  modules: Module[];
  rows: ProgressRow[];
}

export type ListModulesParams = {
trackId?: number;
};

export type ListVideosParams = {
moduleId?: number;
};