# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transform Church Leadership Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List all modules (optionally filter by trackId)
      parameters:
        - name: trackId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a new module (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a single module with its videos and quiz status
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
        "404":
          description: Not found
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update a module (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete a module (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Videos ──────────────────────────────────────────────────────────────
...
  /queue/{videoId}:
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add a video to personal queue
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added to queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueItem"
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove a video from personal queue
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Quizzes ─────────────────────────────────────────────────────────────
...
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/assignments:
    post:
      operationId: createAssignment
      tags: [admin]
      summary: Assign a module to one or more users (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
  /admin/assignments/{assignmentId}:
    delete:
      operationId: deleteAssignment
      tags: [admin]
      summary: Delete an assignment (admin only)
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
      required: [id, clerkId, firstName, lastName, email, role, createdAt]
...
    RoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
      required: [role]
...
    Module:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required: [id, trackId, title, order, createdAt]
    ModuleInput:
      type: object
      properties:
        trackId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
      required: [trackId, title, order]
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
    ModuleDetail:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
        videos:
          type: array
          items:
            $ref: "#/components/schemas/VideoWithProgress"
        quizResult:
          $ref: "#/components/schemas/QuizResult"
        quizUnlocked:
          type: boolean
      required: [id, trackId, title, order, createdAt, videos, quizUnlocked]
    Video:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required: [id, moduleId, title, url, order, createdAt]
...
    VideoWithProgress:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        order:
          type: integer
        createdAt:
          type: string
        progressPercent:
          type: ["integer", "null"]
        completed:
          type: boolean
        inQueue:
          type: boolean
      required: [id, moduleId, title, url, order, createdAt, completed, inQueue]
...
    QueueItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        videoId:
          type: integer
        video:
          $ref: "#/components/schemas/Video"
        addedAt:
          type: string
      required: [id, userId, videoId, video, addedAt]
...
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        moduleId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        passed:
          type: boolean
        takenAt:
          type: string
      required: [id, userId, moduleId, score, totalQuestions, passed, takenAt]
    Assignment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        moduleId:
          type: integer
        module:
          $ref: "#/components/schemas/Module"
        assignedBy:
          type: string
        assignedAt:
          type: string
        dueDate:
          type: ["string", "null"]
        quizResult:
          $ref: "#/components/schemas/QuizResult"
      required: [id, userId, moduleId, module, assignedBy, assignedAt]
    AssignmentInput:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: string
        moduleId:
          type: integer
        dueDate:
          type: ["string", "null"]
      required: [userIds, moduleId]
...