# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transform Church Leadership Training Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  moduleId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  order: number;
  createdAt: string;
}